+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `description` parameter to `Beaker.dataset.create()`, `Beaker.image.create()`, `Beaker.workspace.create()`.
+- Added `public` parameter to `Beaker.workspace.create()`.
+
+### Changed
+
+- `Beaker.workspace.grant_permissions` now takes a `Permission` enum type instead of a raw string, but raw strings will still work.
 
 ## [v0.13.2](https://github.com/allenai/beaker-py/releases/tag/v0.13.2) - 2022-04-29