+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.14.0](https://github.com/allenai/beaker-py/releases/tag/v0.14.0) - 2022-05-04
+
 ### Added
 
 - Added `description` parameter to `Beaker.dataset.create()`, `Beaker.image.create()`, `Beaker.workspace.create()`.